tase-bot-backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
motor>=3.3.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]